requirements.txt
streamlit>=1.28.0
langchain>=0.3.27
langchain-groq>=0.3.8
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.31